refactor(db): move database files in /lib/db
- Updated import paths for AgentDTO and prisma to reflect new directory structure under `db`.
- Moved AgentDTO and prisma file to enhance code clarity and maintainability.
